feat: resolved a security vulnerability within version-control-system
Previously we had the repository as Private; however, for the sake of sharing the repository with CS 6348 faculty, we had to set it to Public. This caused secret values that we had previously been storing within the repository to become comprimised. With this, we have prompty changed the secret values that were comprimised/affected, and prevented future secrets from being shared by blacklisting the .env file which contains said secrets. This has helped to harden security within the application despite an initial setback of needing to revoke prior secrets.